### 2023-04-17 #5381 @editorialbot
[PRE REVIEW]: Reference Point Calculator (RPC): A Shiny app for exploration of limit reference points for fisheries management
**Submitting author:** @quang-huynh (<a href="http://orcid.org/0000-0001-7835-4376">Quang C. Huynh</a>)
**Repository:** https://github.com/Blue-Matter/RPC
**Branch with paper.md** (empty if default branch): JOSS
**Version:** 0.4.0
