### 2023-04-18 50d5be3
Update paper.bib
Added DOIs for JOSS paper bib file
