+++ joss/paper.bib
@@ -32,6 +32,7 @@
   Title = {ICES fisheries management reference points for category 1 and 2 stocks},
   Year = 2021,
   series = {ICES Advice 2021, Section 16.4.3.1.},
+  doi = {10.17895/ices.advice.7891},
   url = {https://doi.org/10.17895/ices.advice.7891}
 }
 
@@ -50,6 +51,7 @@
   Journal = {Fish and Fisheries},
   Title = {Method evaluation and risk assessment: A framework for evaluating management strategies for data-limited fisheries},
   Year = 2023,
+  doi = {10.1111/faf.12726},
   url = {http://doi.org/10.1111/faf.12726}
 }
 
