### 2023-04-18 #5381 @Kevin-Mattheus-Moerman
@quang-huynh the URL is fine but it is not labelled in the bib file as a doi. I attempted a fix for you here: https://github.com/Blue-Matter/RPC/pull/28
