### 2023-04-25 #5381 @graciellehigino
Hi @ivnols and @katherinehebert!

Thank you for accepting to review this paper for JOSS! I am just waiting for the response to one invitation I've sent in these past few days so we can start the review. Meanwhile, feel free to take a look at our review guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) and do not hesitate to contact me if you have any concerns, questions or need help with anything.

I should start the review thread no later than next Monday, May 1st.
