### 2023-05-02 Blue-Matter/RPC#29 @katherinehebert
JOSS review #5418: Manual tests failing for some OMs
Hi! I'm reviewing your submitted JOSS paper at https://github.com/openjournals/joss-reviews/issues/5418

As part of the review, I'm running the 3 manual tests and some of them are failing.

`test-LRP-hist.R` returns:
`[ FAIL 12 | WARN 55 | SKIP 0 | PASS 429 ]`

`test-MP.R` returns:
`[ FAIL 1 | WARN 6 | SKIP 0 | PASS 165 ]`

`test-MSE.R` returns :
`[ FAIL 86 | WARN 6 | SKIP 0 | PASS 341 ]`

The failed tests return the same backtrace:

```
Backtrace:
  1. get(OMs[i]) %>% SubCpars(1:3) %>% ...
  2. MSEtool::runMSE(., Hist = TRUE, silent = TRUE)
  3. MSEtool::Simulate(OM, parallel, silent)
  4. base::data.frame(ObsPars[ind])
  6. base::as.data.frame.list(x[[i]], optional = TRUE, stringsAsFactors = stringsAsFactors)
  8. base (local) `<fn>`(...)
 10. base::as.data.frame.list(x[[i]], optional = TRUE, stringsAsFactors = stringsAsFactors)
 12. base (local) `<fn>`(...)
```
So, it looks like this line is failing for some of the OMs:
`x <- get(OMs[i]) %>% SubCpars(1:3) %>% runMSE(Hist = TRUE, silent = TRUE)`

The Shiny app is otherwise functional when I explore it, so I am unsure whether these failures are critical - but if the test failures could be addressed, it would confirm that the app is fully functional.
