### 2023-05-03 4921e80
Remove un-used arguments in LRP_R() test
