+++ tests/manual/test-LRP-hist.R
@@ -3,12 +3,11 @@
 
 OMs <- avail("OM", package = "RPC")
 
-
-
 for(i in 1:length(OMs)) {
 #for(i in 1:2) {
 
-  x <- get(OMs[i]) %>% SubCpars(1:3) %>% runMSE(Hist = TRUE, silent = TRUE)
+  OM_test <- get(OMs[i])
+  x <- OM_test %>% SubCpars(1:3) %>% runMSE(Hist = TRUE, silent = TRUE)
 
   testthat::expect_s4_class(x, "Hist")
 
@@ -138,18 +137,18 @@ for(i in 1:length(OMs)) {
   testthat::expect_no_error(LRP_R(
     x,
     figure = c("ts"),
-    SR_xlim,
-    SR_ylim,
-    SR_y_RPS0,
+    #SR_xlim,
+    #SR_ylim,
+    #SR_y_RPS0,
     SR_include = 1:3
   ))
 
   testthat::expect_no_error(LRP_R(
     x,
     figure = c("SR"),
-    SR_xlim,
-    SR_ylim,
-    SR_y_RPS0,
+    #SR_xlim,
+    #SR_ylim,
+    #SR_y_RPS0,
     SR_include = 1:3
   ))
 
