### 2023-05-03 71d20dc
runMSE(extended = TRUE) for surplus production plots
