### 2023-05-03 e82d7a7
geom_line argument with linewidth
