+++ R/plot_proj.R
@@ -88,7 +88,8 @@ proj_plot<-function(x, MSEhist, type = c("SSB0", "SSBMSY", "SP", "F", "SPR", "Ca
       structure(dimnames = list(MP = MSEproj@MPs, Year = c(CurrentYr, py[-length(py)]))) %>%
       reshape2::melt()
 
-    g <- ggplot(dat_MP, aes(Year, value)) + geom_line(size = 1, aes(colour = MP)) +
+    g <- ggplot(dat_MP, aes(Year, value)) +
+      geom_line(linewidth = 1, aes(colour = MP)) +
       theme_bw() +
       coord_cartesian(xlim = range(ay)) +
       geom_vline(xintercept = CurrentYr, linetype = 4) +
@@ -100,7 +101,8 @@ proj_plot<-function(x, MSEhist, type = c("SSB0", "SSBMSY", "SP", "F", "SPR", "Ca
       structure(dimnames = list(MP = MSEproj@MPs, Year = c(CurrentYr, py))) %>%
       reshape2::melt()
 
-    g <- ggplot(dat_MP, aes(Year, value)) + geom_line(size = 1, aes(colour = MP)) +
+    g <- ggplot(dat_MP, aes(Year, value)) +
+      geom_line(linewidth = 1, aes(colour = MP)) +
       theme_bw() +
       coord_cartesian(xlim = range(ay), ylim = c(0, 1.2 * max_y)) +
       geom_vline(xintercept = CurrentYr, linetype = 4) +
@@ -125,7 +127,8 @@ proj_plot<-function(x, MSEhist, type = c("SSB0", "SSBMSY", "SP", "F", "SPR", "Ca
     SSB_out$Type <- factor(SSB_out$Type,
                            levels = c("Historical~SSB", "Equilibrium~SSB[0]", "Initial~SSB[0]", "Dynamic~SSB[0]"))
 
-    g <- g + geom_line(data = SSB_out, aes(linetype = Type, size = Type)) +
+    g <- g +
+      geom_line(data = SSB_out, aes(linetype = Type, linewidth = Type)) +
       geom_point(data = SSB_out, aes(shape = Type)) +
       scale_linetype_manual(name = "SSB Type", values = c(1, 2, 1, 3), labels = scales::label_parse()) +
       scale_shape_manual(name = "SSB Type", values = c(NA_integer_, 1, 16, 4), labels = scales::label_parse()) +
@@ -141,7 +144,8 @@ proj_plot<-function(x, MSEhist, type = c("SSB0", "SSBMSY", "SP", "F", "SPR", "Ca
                          Year = ay)
     SSB_out <- rbind(SSBhist, SSBMSY)
 
-    g <- g + geom_line(data = SSB_out, aes(linetype = Type, size = Type)) +
+    g <- g +
+      geom_line(data = SSB_out, aes(linetype = Type, linewidth = Type)) +
       geom_point(data = SSB_out, aes(shape = Type)) +
       scale_linetype_manual(name = "SSB Type", values = c(1, 3), labels = scales::label_parse()) +
       scale_shape_manual(name = "SSB Type", values = c(NA_integer_, 4), labels = scales::label_parse()) +
@@ -149,7 +153,7 @@ proj_plot<-function(x, MSEhist, type = c("SSB0", "SSBMSY", "SP", "F", "SPR", "Ca
 
   } else if(type == "SP") {
     SP <- data.frame(value = hist, Year = hy)
-    g <- g + geom_line(data = SP, size = 2)
+    g <- g + geom_line(data = SP, linewidth = 2)
   } else if(type == "F") {
     FMSY <- data.frame(value = apply(MSEproj@RefPoint$ByYear$FMSY, 2, median),
                        Type = "F[MSY]",
[…]
