### 2023-05-03 9dd6084
Internal max function does not return infinite values
