### 2023-05-03 Blue-Matter/RPC#29 @quang-huynh
Package and test updates for plotting functions and error checking removed the failures on my end

```
remotes::install_github("Blue-Matter/MSEtool", ref = "dev")
remotes::install_github("Blue-Matter/RPC")
```
