### 2023-05-11 56de3a4
Some hist functions plot 25 simulations; closes #27
