### 2023-06-06 b02d88c
Update paper.bib with DOI
