### 2023-06-06 #5418 @quang-huynh
@graciellehigino The DOIs should be properly specified now. Yes the release version for the journal will be 0.4.1
